Renamed the application to botManager
