+++ botManager/.gitignore
