+++ botManager/README.md
@@ -1,0 +1,14 @@
+# botManager
+
+## How to use
+
+To run this application you'll need [Node.js](https://nodejs.org/en/download/) (and [npm](http://npmjs.com), which is included in Node.js) installed on your computer. From your command line:
+
+```bash
+# Go to the codeTest folder you've cloned
+cd botManager
+# Install dependencies
+npm install
+# Run the app!
+npm start
+```
