+++ botManager/fonts/glyphicons-halflings-regular.ttf
