+++ botManager/fonts/glyphicons-halflings-regular.woff
