+++ botManager/js/bootstrap.js
