+++ botManager/js/npm.js
