--- codeTest/README.md
@@ -1,14 +1,0 @@
-# Slither.io-bot code test
-
-## How to use
-
-To run this application you'll need [Node.js](https://nodejs.org/en/download/) (and [npm](http://npmjs.com), which is included in Node.js) installed on your computer. From your command line:
-
-```bash
-# Go to the codeTest folder you've cloned
-cd codeTest
-# Install dependencies
-npm install
-# Run the app!
-npm start
-```
