--- codeTest/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "slitherio-code-test",
-  "version": "0.0.1",
-  "description": "An application to test code for a slither.io-bot",
-  "main": "main.js",
-  "scripts": {
-    "start": "electron main.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ErmiyaEskandary/Slither.io-bot.git"
-  },
-  "keywords": [
-    "slither.io",
-    "bot",
-    "code",
-    "test"
-  ],
-  "author": "K00sKlust",
-  "license": "MPL-2.0",
-  "bugs": {
-    "url": "https://github.com/ErmiyaEskandary/Slither.io-bot/issues"
-  },
-  "homepage": "https://github.com/ErmiyaEskandary/Slither.io-bot#readme",
-  "devDependencies": {
-    "electron-prebuilt": "^1.0.1"
-  }
-}
